requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.2.0
certifi==2024.2.2
click==8.1.7
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
